requirements.txt
numpy==1.20.2
opencv_contrib_python==4.5.4.60
Pillow==9.0.1
PyYAML==6.0
torch==1.8.0
torchvision==0.9.1
xlwt==1.3.0
